correct confirmParticipation

+++ main-server/src/main/java/com/example/mainserver/participation/model/ParticipationChangeStatus.java
@@ -11,10 +11,10 @@ import java.util.List;
 
 public class ParticipationChangeStatus {
 
-    private List<Integer> requestIds;
+    private List<Long> requestIds;
     private final StatusRequest status;
 
-    public ParticipationChangeStatus(List<Integer> requestIds, StatusRequest statusRequest) {
+    public ParticipationChangeStatus(List<Long> requestIds, StatusRequest statusRequest) {
         this.requestIds = requestIds;
         this.status = statusRequest;
     }
